requirements.txt
# MCP SDK - Official Model Context Protocol Python SDK
mcp>=1.8.0

# Web framework (FastMCP dependency)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Web scraping libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.6.0
newspaper3k>=0.2.8

# Search engine
duckduckgo-search>=4.1.1

# Utilities
python-dotenv>=1.0.0